Store images and thumbnails in DB
I know this usually a bad idea, but I have nowhere else to put them.
This fixes the controller tests, and adds a custom Mix task for updating
the current images. In the future, I'd like to offload this work so the
image fetching isn't done in request on create. That's a bad setup
currently, but this is the first try!